+++ lib/api/injection.dart
@@ -1,30 +1,10 @@
 import 'package:dio/dio.dart';
-import 'package:flutter/foundation.dart';
-import 'package:pretty_dio_logger/pretty_dio_logger.dart';
-
 import 'api_client.dart';
 
 final apiClient = RestClient(setDio());
 
 Dio setDio() {
   final dio = Dio();
-  dio.options.baseUrl = 'https://api.daakia.co.in/v2.0/';
-  dio.interceptors.add(PrettyDioLogger(
-      requestHeader: true,
-      requestBody: true,
-      responseBody: true,
-      responseHeader: false,
-      error: true,
-      compact: true,
-      maxWidth: 90,
-      enabled: kDebugMode,
-      filter: (options, args) {
-        // don't print requests with uris containing '/posts'
-        if (options.path.contains('/posts')) {
-          return false;
-        }
-        // don't print responses with unit8 list data
-        return !args.isResponse || !args.hasUint8ListData;
-      }));
+  dio.options.baseUrl = 'https://stag-api.daakia.co.in/v2.0/';
   return dio;
 }